🐞[fix] hotfix
- sse heartbeat 전송 분리. connection 문제 해결
